+++ python/GafferSceneUI/SceneViewUI.py
@@ -66,6 +66,11 @@ def __rendererPlugActivator( plug ) :
 
 	return plug.parent()["name"].getValue().lower() == plug.getName().lower()
 
+def __condensedEditScopeSpacerActivator( node ) :
+
+	input = node["editScope"].getInput()
+	return input is not None and input.getName() == "editScope" and isinstance( input.node(), GafferUI.Editor.Settings )
+
 Gaffer.Metadata.registerNode(
 
 	GafferSceneUI.SceneView,
@@ -92,6 +97,12 @@ Gaffer.Metadata.registerNode(
 	"toolbarLayout:customWidget:RightEditScopeBalancingSpacer:widgetType", "GafferSceneUI.SceneViewUI._RightEditScopeBalancingSpacer",
 	"toolbarLayout:customWidget:RightEditScopeBalancingSpacer:section", "Top",
 	"toolbarLayout:customWidget:RightEditScopeBalancingSpacer:index", -2,
+
+	"toolbarLayout:activator:condensedEditScopeMenu", __condensedEditScopeSpacerActivator,
+	"toolbarLayout:customWidget:CondensedEditScopeBalancingSpacer:widgetType", "GafferSceneUI.SceneViewUI._CondensedEditScopeBalancingSpacer",
+	"toolbarLayout:customWidget:CondensedEditScopeBalancingSpacer:section", "Top",
+	"toolbarLayout:customWidget:CondensedEditScopeBalancingSpacer:index", -2,
+	"toolbarLayout:customWidget:CondensedEditScopeBalancingSpacer:visibilityActivator", "condensedEditScopeMenu",
 
 	"nodeToolbar:right:type", "GafferUI.StandardNodeToolbar.right",
 
@@ -1259,6 +1270,17 @@ class _RightEditScopeBalancingSpacer( GafferUI.Spacer ) :
 			preferredSize = imath.V2i( width, 1 ),
 			maximumSize = imath.V2i( width, 1 )
 		)
+
+# This Spacer balance the right side of the toolbar by
+# preserving the width lost when the EditScope menu is
+# displayed in condensed form.
+class _CondensedEditScopeBalancingSpacer( GafferUI.Spacer ) :
+
+	def __init__( self, sceneView, **kw ) :
+
+		editScopeWidth = Gaffer.Metadata.value( sceneView["editScope"], "toolbarLayout:width" ) or 130
+		# EditScope width - spacer - condensed EditScope width
+		width = max( editScopeWidth - 4 - 50, 0 )
 		GafferUI.Spacer.__init__(
 			self,
 			imath.V2i( 0 ), # Minimum